Get window query working The first exception of a continuous query is retained for analysis. We better set the access modes when updating a BAT. Use long numbers to collect a sum.

+++ sql/backends/monet5/datacell/Tests/scenario04.sql
@@ -11,7 +11,7 @@ create table datacell.winin(
     tag timestamp,
     payload integer
 );
-create table datacell.winout( tag timestamp, mi integer, ma integer, su integer);
+create table datacell.winout( tag timestamp, mi integer, ma integer, su bigint);
 
 call datacell.receptor('datacell.winin','localhost',50500);
